+++ src/pytest_subtests/plugin.py
@@ -34,10 +34,7 @@ from _pytest.unittest import TestCaseFunction
 if TYPE_CHECKING:
     from types import TracebackType
 
-    if sys.version_info < (3, 8):
-        from typing_extensions import Literal
-    else:
-        from typing import Literal
+    from typing import Literal
 
 
 def pytest_addoption(parser: pytest.Parser) -> None: